+++ CHANGELOG.md
@@ -8,7 +8,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 
 ### Added
-- Trigger share event to analytics on social button click 
+- Trigger vtex:share event to analytics on social button click 
+- Tests to vtex:share for all social buttons case
 
 ## [3.167.0] - 2023-03-28